backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.19
python-jose[cryptography]==3.3.0
supabase==2.13.0
httpx==0.28.1
anthropic==0.42.0
python-docx==1.1.2
pdfplumber==0.11.4
openpyxl==3.1.5
